Refactor by dividing code into separate files
The PYTHONPATH for pytest also had to be updated to deal with import
issues after the refactor.
