+++ src/exceptions.py
@@ -1,0 +1,4 @@
+class MTPAnalyzerException(Exception):
+    """Base class for exceptions for this app."""
+
+    pass
