+++ src/version.py
@@ -1,0 +1,3 @@
+__author__ = "Simon"
+__version__ = "0.1.0"
+__licence__ = "MIT"
